Fix for 2170453: Capitalization sticks when entering passwords in Browser
This is caused by a delay in changing the shift state for performance reasons.
The delay was too long, causing multiple characters to be in shifted state when
typing fast (300ms interval).

Fix removes the optimization for the forward typing case. Doesn't affect the
backspace long-press optimization.

Triage: Hiroshi
Dr. No: mcleron