Fix for #2244624 : Keyboard freezes up sometimes
I think the cause for this is the Contacts database being updated. This
causes the keyboard to reload the contacts once every 30 minutes. Since it
loads it synchronously, it affects people with several thousand contacts.

Although in my tests, with 3000 contacts, the delay was only 600ms, I've
had several reports from long-time googlers about this problem, so I'm
switching to loading the contacts asynchronously in a background thread.

Also fix a potential problem with capitalizing "i" if a contact has "i" as
one of the names.