DO NOT MERGE: Fix for 2295810: Auto-correction of English results in Englishman
Comparisons were always happening with lowercase version of the typed
word, which wouldn't match the uppercase word in the dictionary, so it
became an unrecognized word when typed in full. Highlight was then going
to the next word in the list.

Fix compares the lowercase and uppercase versions of the word for
validity.

Merge from eclair-mr2