Save and restore Auto Dictionary data.
Due to the addition of input language switching, the auto dictionary
will be re-created very frequently. We need to save it and restore it
during transitions, otherwise the data will be purged too often to be
of much use in the multi-lingual case.

This also fixes the case where a user frequently turns the phone off/on.
Auto dictionary data was not being saved in those situations either.

The dictionary will probably never grow too large since most of the
frequently used words will be moved to the user dictionary any way.

Also, now the input locale is different from the display locale, so
save the input locale in the UserDictionary class and insert words
into the user dictionary with the correct locale.

And NPE fix for bug: 2464573