Treat fast moves as multi-touch. Bug: 2561786
Detect sudden moves that have a large delta that would not be a typical drag
by the user when hunting for a key. Drop move events until the user releases
the touch and register both starting and ending keys instead of only the
ending key. This reduces the frequency of dropped keys on touch hardware that
isn't able to detect multi-touch reliably.

This also takes into account diagonal moves or right angled moves due to
pseudo-multi-touch.

It weakens the swipe gestures (to swipe, you'll need to swipe slowly and fully,
  otherwise it will trigger 2 keys)

Change-Id: Ieba3a0bef4c8910f0f17aa20baedb3581d1de12b