Disable pop-up preview on mini keyboard
This change also disable the special behavior of long press a key that
has only one digit alternate.  If the visual design of both the pop-up
preview and the mini keyboard will look like same, I think that
showing mini keyboard is enough for the key that has one digit
alternate.

Bug: 2973546
Change-Id: I04f138c120aa58526fe054626dac856171fb7a58