Introduce global shared variables for current keyboard settings.
Yes, globals are evil. But the persisted shared preferences are global data
by definition, and trying to hide this by propagating the current manually
just adds a lot of complication. This is especially annoying due to Views
getting constructed in a way that doesn't support adding additional
constructor arguments, requiring post-construction method calls, which is
error-prone and fragile.

This CL adds a GlobalKeyboardSettings class with public member fields,
accessible as LatinIME.sKeyboardSettings from other classes. Comments in the
file explain who is responsible for writing them, and who is reading them.