New theming logic for recoloring symbols
Add a "recolorSymbols" flag for themes. If it's active, use runtime filtering
to re-draw the existing white icons in a different color on the fly.

Also adds a 'backgroundAlpha' property for use with the experimental transparent keyboard.

In both cases, make sure that the new (and slower) rendering path is only used
for themes that need it, to help avoid performance regressions on slower
hardware.