Add support for padding space around keys
The themes didn't look very nice on tablets since the keys had a very narrow
gutter between them. In theory it would be possible to work around this by
using the existing key gap attributes, but this is a pain since these gaps are
added to the key size, requiring all the key dimensions to be adjusted in the
XML file to make the keys fit.

Instead, add new horizontalPad and verticalPad attributes to keys, these shrink
the key image without changing the overall key geometry as far as the layout
grid is concerned.

This is configured with new key_horizontal_pad and key_vertical_pad dimensions,
replacing the existing key_bottom_gap which is now set to zero.

Note that this changes the sizing of portrait mode keyboards a bit, these used
to be bigger than the configured screen height percentage since the bottom_gap
size was added on top of the expected size. Users may need to tweak the
keyboard height percentage setting if they feel it's now too small.