Refactor modifier key handling
The modifier key handling was fairly broken and only really worked right with
ConnectBot. Other Android applications got key events without the appropriate
SHIFT, CTRL and META modifiers, I had originally removed those due to issue 24
(Shift causes next number to be entered as shift+number).

Now that Honeycomb and ICS actually make use of the Ctrl key, for example for
Ctrl-X/C/V/A shortcuts, need to make sure that the key modifiers get reported
properly in the generated key events.

The new system now generates appropriate low-level key events when modifier
keys are active instead of relying on the input method libraries to do so. This
is based on an ascii-to-keycode map to determine the appropriate base
character. Don't use the odd special handling for digits that had caused the
breakage in issue 24.

TODO: watch out for regressions, key event handling is a bit confusing in
Android, especially in older versions that had odd dependencies on built-in
hardcoded base keymaps.