Reorganize preferences, add input connection info
The settings menu was a bit of a mess. Originally I had just added my new
options on the top and left the Gingerbread options in their original order
below, but this didn't scale well with the rapidly increasing number of new
options.

Recategorize the options into (hopefully) logical categories, including a new
"Debugging" section at the bottom. Add a new item there that shows information
about the current input connection, including the type and flags of the current
input field.