Add compact 5-row portrait layout intended for phones
This layout is 11 columns across, so the keys are a little bit smaller than the default with 10 columns, and significantly bigger than the full 14-column layout.

Also add a hack to disable the completions based on the "suggestions in landscape mode" setting. TODO: this should be its own setting, defer fixing this until mode-specific config override is available.