Fix Turkish alt chars for 4-row mode.
Includes a special case for the popup builder to handle ıiIİ. The 4-row mode has i as main char and I as both shift and caps char for compatibility with US QWERTY keyboards. The alternates include ıi and should be cased and uniqued appropriately.

Normally, the isShifted and isShiftCaps flags represent three states, but this key has the otherwise-unused state where isShifted is false (isSimpleUppercase is false since it considers I to be a distinct different letter and not the uppercase version of i in this locale) and isCapsShift is true. When that happens, the alternates should be upcased.