Disable ContactsDictionary, remove Contacts permission
I'm now convinced that the contacts autocomplete dictionary is more trouble
than it's worth. It requires a suspicious-looking permission, and uses
a ton of memory in some configurations, i.e. if you have a corporate account
synced.

If people feel strongly about this functionality, it should be possible to
write a separate app or service that syncs contact names to the user dictionary.