Add Ctrl-A options, require Ctrl-Alt-A by default
Add an option to configure "Ctrl-A" behavior to work around
accidental text deletion when doing an unintended "select all"
while typing.

New default is to ignore Ctrl-A, and require use of Ctrl-Alt-A
instead to do a "select all". Pressing Ctrl-A shows a toast message
indicating this.

The other available settings are to ignore Ctrl-A completely
(with no toast message), or to enable Ctrl-A normally with no
behavior override.

See issue #487.